# server/internal/app/testdata/probe/openapi.yaml
openapi: 3.1.0
info: { title: The probe module's contract, version: '1' }
paths:
...
  /auth/deletion/cancel:
    post: { operationId: postAuthDeletionCancel, security: [], requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '204': { description: Cancelled }, default: { $ref: '#/components/responses/Problem' } } }
  /me/deletion:
    post: { operationId: postMeDeletion, requestBody: { $ref: '#/components/requestBodies/Any' }, responses: { '202': { description: Scheduled }, default: { $ref: '#/components/responses/Problem' } } }
...
  /me/exports:
    get: { operationId: getMeExports, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
    post: { operationId: postMeExports, responses: { '202': { description: Queued }, default: { $ref: '#/components/responses/Problem' } } }
  /me/exports/{export_id}:
    parameters: [{ name: export_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    get: { operationId: getMeExportsByExportId, responses: { '200': { description: Found }, default: { $ref: '#/components/responses/Problem' } } }
...
  /households/{household_id}/exports:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }]
    get: { operationId: getExports, responses: { '200': { description: Listed }, default: { $ref: '#/components/responses/Problem' } } }
    post: { operationId: postExports, responses: { '202': { description: Queued }, default: { $ref: '#/components/responses/Problem' } } }
  /households/{household_id}/exports/{export_id}:
    parameters: [{ $ref: '#/components/parameters/HouseholdId' }, { name: export_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }]
    get: { operationId: getExportsByExportId, responses: { '200': { description: Found }, default: { $ref: '#/components/responses/Problem' } } }
...
components:
  parameters:
    HouseholdId: { name: household_id, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
  requestBodies:
    Any: { required: true, content: { application/json: { schema: { type: object } } } }
...
  responses:
    Problem:
      description: A problem.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Uuid:
      type: string
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
...
    ProblemCode: { type: string }
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { $ref: '#/components/schemas/ProblemCode' }
        request_id: { type: string }
...